modularized CopyLinkButton further
